place toDoList component into seperate js file
