--- src/Todo.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Todo (props) {
-    return (
-        <div className = 'task-card'>
-        <p>{props.todo.task}</p>
-        </div>
-    )
-}
-
-export default Todo;
